refactor: simplify CLI flag names and enhance mock relay implementation for testing
